edited roots3and4 so it uses floats.

+++ Roots3And4.cpp
@@ -35,13 +35,13 @@
 #define	    IsZero(x)	((x) > -EQN_EPS && (x) < EQN_EPS)
 
 #ifdef NOCBRT
-#define     cbrt(x)     ((x) > 0.0 ? pow((double)(x), 1.0/3.0) : \
-                          ((x) < 0.0 ? -pow((double)-(x), 1.0/3.0) : 0.0))
+#define     cbrt(x)     ((x) > 0.0 ? pow((float)(x), 1.0/3.0) : \
+                          ((x) < 0.0 ? -pow((float)-(x), 1.0/3.0) : 0.0))
 #endif
 
-int SolveQuadric(double c[3], double s[2])
+int SolveQuadric(float c[3], float s[2])
 {
-    double p, q, D;
+    float p, q, D;
 
     /* normal form: x^2 + px + q = 0 */
 
@@ -61,7 +61,7 @@ int SolveQuadric(double c[3], double s[2])
     }
     else /* if (D > 0) */
     {
-    double sqrt_D = sqrt(D);
+    float sqrt_D = sqrt(D);
 
     s[ 0 ] =   sqrt_D - p;
     s[ 1 ] = - sqrt_D - p;
@@ -70,13 +70,13 @@ int SolveQuadric(double c[3], double s[2])
 }
 
 
-int SolveCubic(double c[4], double s[3])
+int SolveCubic(float c[4], float s[3])
 {
     int     i, num;
-    double  sub;
-    double  A, B, C;
-    double  sq_A, p, q;
-    double  cb_p, D;
+    float  sub;
+    float  A, B, C;
+    float  sq_A, p, q;
+    float  cb_p, D;
 
     /* normal form: x^3 + Ax^2 + Bx + C = 0 */
 
@@ -103,9 +103,9 @@ int SolveCubic(double c[4], double s[3])
         s[ 0 ] = 0;
         num = 1;
     }
-    else /* one single and one double solution */
+    else /* one single and one float solution */
     {
-        double u = cbrt(-q);
+        float u = cbrt(-q);
         s[ 0 ] = 2 * u;
         s[ 1 ] = - u;
         num = 2;
@@ -113,8 +113,8 @@ int SolveCubic(double c[4], double s[3])
     }
     else if (D < 0) /* Casus irreducibilis: three real solutions */
     {
-    double phi = 1.0/3 * acos(-q / sqrt(-cb_p));
-    double t = 2 * sqrt(-p);
+    float phi = 1.0/3 * acos(-q / sqrt(-cb_p));
+    float t = 2 * sqrt(-p);
 
     s[ 0 ] =   t * cos(phi);
     s[ 1 ] = - t * cos(phi + M_PI / 3);
@@ -123,9 +123,9 @@ int SolveCubic(double c[4], double s[3])
     }
     else /* one real solution */
     {
-    double sqrt_D = sqrt(D);
-    double u = cbrt(sqrt_D - q);
-    double v = - cbrt(sqrt_D + q);
+    float sqrt_D = sqrt(D);
+    float u = cbrt(sqrt_D - q);
+    float v = - cbrt(sqrt_D + q);
 
     s[ 0 ] = u + v;
     num = 1;
@@ -142,12 +142,12 @@ int SolveCubic(double c[4], double s[3])
 }
 
 
-int SolveQuartic(double c[5], double s[4])
+int SolveQuartic(float c[5], float s[4])
 {
-    double  coeffs[ 4 ];
-    double  z, u, v, sub;
-    double  A, B, C, D;
-    double  sq_A, p, q, r;
+    float  coeffs[ 4 ];
+    float  z, u, v, sub;
+    float  A, B, C, D;
+    float  sq_A, p, q, r;
     int     i, num;
 
     /* normal form: x^4 + Ax^3 + Bx^2 + Cx + D = 0 */
