+++ Roots3And4.h
@@ -8,5 +8,5 @@
 
 #ifndef H_ROOTS3AND4
 #define H_ROOTS3AND4
-int SolveQuartic(double c[5], double s[4]);
+int SolveQuartic(float c[5], float s[4]);
 #endif //!H_ROOTS3AND4
